Handle mustaches with hash arguments
Because racc is a LALR(1) parser generator, there is not enough
lookahead to distinguish an ID that is an expression from an ID that is
part of a hash segment.

To work around is, detection of the key part of the hash segment is
moved to the lexer. This means the lexer now recognizes the ID and the
equals sign together as one token.

This change resolves the last remaining shift/reduce conflict in the
parser.