Fix shift/reduce conflicts
The old rules allowed any number of 'none' items in a row to be returned
for contents. This change separates out the empty case for contents just
like with exprs.

It also changes statement back to allow just a single content per
statement. To keep things working with the change in parsing, whitespace
handling needs to be adjusted to combine adjacent :content expressions.