Add unique node type at the root of the AST
This is needed to figure out if a statement is the first element in the
parsed handlebars template. In that position, a block opener element
only needs to be followed by whitespace and a line break to be
considered standalone. In all other positions, it must alse be preceded
by a line break and whitespace.