Split whitespace handler into two parts
This splits WhitespaceHandler into a part for handling 'standalone'
blocks and partials (StandaloneWhitespaceHandler), and a part for
handling the stripping decorations on mustaches (WhitespaceStripper).

This means the handling of the :ignore_standalone compile-time option
can be done entirely inside the Parser class, since ignoring standalone
means simply not calling the StandaloneWhitespaceHandler.