Strip whitespace around standalone partial
A partial may expand to several lines. If the partial is used
standalone, i.e., it is on a line by itself with only whitespace before
and after it, the whitespace should not be rendered outside of the
partial. Instead, the rendering of the partial itself is responsible for
indenting based on the whitespace originally present on the line before
the partial.

To make this work, the first step is to remove the whitespace that is
present on the line before (and after) the standalone partial.

To make this possible, this commit separates the collection of
information on whitespace from the application of that information.