Handle standalone partials on the first line of the template
Adding support for standalone partials exposed an issue in this
implementation: Comparing s-expressions compares by value, not by
object_id. This meant that the checks comparing item and before to the
first or last statement could incorrectly pass. This commit therefore
also changes the code to check by index instead of by value of the
items.