Handle standalone partials on the last line of the template
This also fixes an issue due to an inconsistent position of the strip
flags caused by the recent addition of the indent sub-expression. The
location of this indent expression was not consistent, causing it to be
interpreted as a strip expression in some cases.

To resolve this, an empty position is added to the partial expression,
that can hold the indent expression when an indent is set. In this way,
the positions of the sub-expressions of the partial expression are once
again the same in all cases.